Fix
Fix

+++ Debian/1-pre-install.sh
@@ -58,13 +58,14 @@ elif [ $r == 2 ]; then
 
     echo -e "deb http://mxrepo.com/mx/repo/ $deb_cn ahs main non-free" | sudo tee -a /etc/apt/sources.list.d/mx.list
 
-    curl https://mxrepo.com/mx27repo.asc | apt-key add -
+    curl -s https://mxrepo.com/mx27repo.asc | apt-key add -
+    if test -f "/etc/apt/trusted.gpg"; then
     mv /etc/apt/trusted.gpg /etc/apt/mx.gpg
     ln -s /etc/apt/mx.gpg /etc/apt/trusted.gpg.d/mx.gpg
     sleep 5
     echo " "
     echo " "
-
+    fi
     curl -s https://mxrepo.com/mx25repo.asc | apt-key add -
     if test -f "/etc/apt/trusted.gpg"; then
     mv /etc/apt/trusted.gpg /etc/apt/mx.gpg
